refactor(s3)!: extract the S3 uploader into a dial9-s3 crate
Moves the S3 uploader plus its circuit breaker and instance metadata into a new dial9-s3 crate with a flat API
(dial9_s3::S3Config).
