+++ .github/workflows/ci.yml
@@ -345,7 +345,9 @@ jobs:
             -p dial9-trace-format \
             -p dial9-perf-self-profile \
             -p dial9-macro \
-            -p dial9-tokio-telemetry
+            -p dial9-tokio-telemetry \
+            -p dial9-metrique \
+            -p dial9-s3
 
   package:
     name: Cargo package
@@ -368,8 +370,10 @@ jobs:
             -p dial9-trace-format \
             -p dial9-core \
             -p dial9-utils \
+            -p dial9-s3 \
             -p dial9-perf-self-profile \
             -p dial9-macro \
+            -p dial9-metrique \
             -p dial9-tokio-telemetry \
             -p dial9-viewer
