+++ dial9-s3/src/instance_metadata.rs
