+++ dial9-tokio-telemetry/tests/s3_integration.rs
@@ -7,7 +7,7 @@ mod fake_s3;
 use aws_config::Region;
 use aws_sdk_s3::Client;
 use common::{fast_sealing_writer, wait_for_sealed_segment};
-use dial9_tokio_telemetry::background_task::s3::S3Config;
+use dial9_s3::S3Config;
 use dial9_tokio_telemetry::telemetry::{
     DiskBuffer, RecorderPipelineExt, RecorderS3ClientExt, TokioAttachOptions, recorder, spawn,
 };
