+++ dial9-utils/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.5.0-rc0"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
-description = "Pipeline and runtime integration utilities for dial9"
+description = "Opt-in integrations for dial9"
 
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ["cfg(shuttle)"] }
@@ -13,8 +13,7 @@ missing_debug_implementations = "warn"
 [dependencies]
 dial9-core = { workspace = true, features = ["pipeline"] }
 dial9-trace-format = { workspace = true, optional = true }
-bon = "3"
-tokio = { version = "1", features = ["fs"], default-features = false }
+tokio = { version = "1", default-features = false }
 tracing = "0.1.44"
 axum-0-7 = { package = "axum", version = "0.7", optional = true }
 axum-core-0-4 = { package = "axum-core", version = "0.4", optional = true }
@@ -28,17 +27,6 @@ hyper-util = { version = "0.1", features = [
 ], optional = true }
 tower = { version = "0.5", optional = true }
 tower-service = { version = "0.3", optional = true }
-aws-sdk-s3 = { version = "1", default-features = false, features = [
-    "behavior-version-latest",
-    "sigv4a",
-    "rt-tokio",
-    "default-https-client",
-] }
-aws-config = { version = "1", default-features = false }
-time = { version = "0.3", features = ["formatting", "macros"] }
-hostname = "0.4"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
 
 [features]
 ## A traced replacement for `axum::serve` from Axum 0.7.
@@ -69,21 +57,13 @@ axum-08 = [
 ]
 
 [dev-dependencies]
-dial9-core = { workspace = true, features = ["pipeline", "test-util"] }
-assert2 = { workspace = true }
-async-trait = "0.1.89"
-flate2 = "1"
+dial9-core = { workspace = true, features = ["pipeline"] }
 tempfile = "3"
 tokio = { version = "1.52.0", features = [
     "rt",
     "rt-multi-thread",
     "macros",
-    "io-util",
     "net",
     "sync",
     "time",
-    "test-util",
 ] }
-s3s = "0.14.1"
-s3s-fs = "0.14.1"
-s3s-aws = "0.14.1"
