+++ dial9/Cargo.toml
@@ -40,6 +40,7 @@ dial9-core = { workspace = true }
 dial9-trace-format = { workspace = true }
 dial9-perf-self-profile = { workspace = true, optional = true }
 dial9-tokio-telemetry = { version = "0.5.0-rc1", path = "../dial9-tokio-telemetry", optional = true }
+dial9-s3 = { workspace = true, optional = true }
 dial9-macro = { workspace = true, optional = true }
 dial9-metrique = { workspace = true, optional = true }
 
@@ -126,7 +127,7 @@ deserialize = ["dial9-trace-format/serde-deserialize"]
 analysis = ["deserialize", "tokio", "dial9-tokio-telemetry/analysis"]
 
 ## S3 upload pipeline stage.
-worker-s3 = ["tokio", "dial9-tokio-telemetry/worker-s3"]
+worker-s3 = ["tokio", "dep:dial9-s3", "dial9-tokio-telemetry/worker-s3"]
 
 ## Async backtraces at Tokio yield points (Linux only).
 taskdump = ["tokio", "dial9-tokio-telemetry/taskdump"]
