+++ release-plz.toml
@@ -37,6 +37,14 @@ name = "dial9-utils"
 changelog_update = false
 
 [[package]]
+name = "dial9-s3"
+changelog_update = false
+
+[[package]]
+name = "dial9-metrique"
+changelog_update = false
+
+[[package]]
 name = "dial9-macro"
 changelog_update = false
 
@@ -68,6 +76,8 @@ changelog_path = "CHANGELOG.md"
 changelog_include = [
     "dial9-core",
     "dial9-utils",
+    "dial9-s3",
+    "dial9-metrique",
     "dial9-macro",
     "dial9-perf-self-profile",
     "dial9-trace-format",
